+++ src/components/SurveyStep3.tsx
@@ -1,11 +1,12 @@
 import styled from "styled-components";
-import SelectButton from "./SelectButton";
+import { Link } from "react-router-dom";
+import colors from "../styles/colors";
+import PersonalColorButton from "./PersonalColorButton";
 import { useState } from "react";
 
 const Wrapper = styled.div`
   margin-top: 1rem;
   width: 100%;
-  padding-right: 1rem;
 `;
 
 const TitleWrapper = styled.div`
@@ -17,91 +18,65 @@ const TitleWrapper = styled.div`
 
 const AnswerWrapper = styled.div`
   display: grid;
-  grid-template-columns: repeat(3, 1fr); // 한 줄에 2개
+  grid-template-columns: repeat(2, 1fr); // 한 줄에 2개
   gap: 0.75rem;
   margin-top: 2rem;
+`;
+
+const SkinTestWrapper = styled.div`
+  display: flex;
+  margin-top: 3rem;
+  font-size: 0.75rem;
+  color: ${colors.mediumGrey};
+  text-decoration: underline;
+  justify-content: center;
 `;
 
 const Options = [
   {
     id: 1,
-    name: "잡티",
+    name: "봄 웜",
+    colors: ["#FFD8A9", "#FFC0A9", "#FFE6CC"],
   },
   {
     id: 2,
-    name: "미백",
+    name: "여름 쿨",
+    colors: ["#C3D5FF", "#E0B5DB", "#A4B8D1"],
   },
   {
     id: 3,
-    name: "주름",
+    name: "가을 웜",
+    colors: ["#CFA36E", "#A06045", "#E2C1A0"],
   },
   {
     id: 4,
-    name: "모공",
-  },
-  {
-    id: 5,
-    name: "탄력",
-  },
-  {
-    id: 6,
-    name: "홍조",
-  },
-  {
-    id: 7,
-    name: "각질",
-  },
-  {
-    id: 8,
-    name: "트러블",
-  },
-  {
-    id: 9,
-    name: "블렉헤드",
-  },
-  {
-    id: 10,
-    name: "피지과다",
-  },
-  {
-    id: 11,
-    name: "민감성",
-  },
-  {
-    id: 12,
-    name: "아토피",
-  },
-  {
-    id: 13,
-    name: "다크서클",
+    name: "겨울 쿨",
+    colors: ["#C5CBE1", "#B4A4DD", "#333C57"],
   },
 ];
 
-export default function SurveyStep1() {
-  const [selected, setSelected] = useState<String[]>([]);
-
-  const toggleSelection = (name: string) => {
-    if (selected?.includes(name)) {
-      setSelected(selected.filter((item) => item !== name));
-    } else {
-      setSelected([...selected, name]);
-    }
-  };
+export default function SurveyStep2() {
+  const [selected, setSelected] = useState<String>();
 
   return (
     <>
       <Wrapper>
-        <TitleWrapper>피부 고민이 있으신가요?</TitleWrapper>
+        <TitleWrapper>퍼스널 컬러에 대해 알려주세요!</TitleWrapper>
         <AnswerWrapper>
           {Options.map((item) => (
-            <SelectButton
+            <PersonalColorButton
               buttonName={item.name}
-              size="small"
-              isActivated={selected.includes(item.name)}
-              onClick={() => toggleSelection(item.name)}
+              isActivated={selected === item.name}
+              colors={item.colors}
+              onClick={() => setSelected(item.name)}
             />
           ))}
         </AnswerWrapper>
+        <Link to="">
+          <SkinTestWrapper>
+            잘 모르겠다면? 퍼스널 컬러 진단을 받아보세요
+          </SkinTestWrapper>
+        </Link>
       </Wrapper>
     </>
   );
